build: Fix compilation issue for non-linux platform
This commit is to make sure there is no compilation issue
with windows for netlink and netns packages

Relates to https://github.com/vishvananda/netlink/issues/146
Relates to https://github.com/vishvananda/netns/issues/23

Suggested-by: Tobias Klauser <tobias@cilium.io>
